# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Capalyst QMS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      parameters:
        - in: query
          name: page
          schema:
            type: ["integer", "null"]
        - in: query
          name: pageSize
          schema:
            type: ["integer", "null"]
        - in: query
          name: search
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"

    post:
      operationId: createUser
      tags: [users]
      summary: Create / add a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user (Admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "403":
          description: Forbidden
        "404":
          description: Not found
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user role and department
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /user-role-definitions:
    get:
      operationId: listUserRoleDefinitions
      tags: [users]
      summary: List all defined user roles (built-in + custom)
      responses:
        "200":
          description: Role definitions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserRoleDefinition"
    post:
      operationId: createUserRoleDefinition
      tags: [users]
      summary: Create a new custom user role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRoleDefinitionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRoleDefinition"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, roles, status, mustChangePassword]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        roles:
          type: array
          items:
            type: string
        department:
          type: ["string", "null"]
        status:
          type: string
        mustChangePassword:
          type: boolean
    UserListResponse:
      type: object
      required: [data, total, page, pageSize]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
    UserInput:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: ["string", "null"]
        roles:
          type: array
          items:
            type: string
        department:
          type: ["string", "null"]
    UserUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        roles:
          type: array
          items:
            type: string
        department:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        password:
          type: ["string", "null"]
        mustChangePassword:
          type: ["boolean", "null"]
    UserRoleDefinition:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
    UserRoleDefinitionInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
...